src/Logistic/requirements.txt
numpy
joblib
typing
<<<<<<< HEAD
pandas
nltk
regex
scikit-learn
tqdm
python-dotenv
=======
>>>>>>> 9643c8dbf8c807e406368a719a7a61d4b9b6399e
